requirements.txt
flask
pymongo
gunicorn
flask-cors
python-dotenv
jsonschema